+++ jdbc/src/main/java/org/apache/metamodel/jdbc/JdbcCreateTableBuilder.java
@@ -93,23 +93,22 @@ final class JdbcCreateTableBuilder extends AbstractTableCreationBuilder<JdbcUpda
 			sb.append(getUpdateCallback().quoteIfNescesary(column.getName()));
 			sb.append(' ');
 			final String nativeType = column.getNativeType();
+			final Integer columnSize = column.getColumnSize();
 			if (nativeType == null) {
 				ColumnType columnType = column.getType();
 				if (columnType == null) {
 					columnType = ColumnType.VARCHAR;
 				}
 
-				final String columnTypeString = queryRewriter.rewriteColumnType(columnType);
-
+				final String columnTypeString = queryRewriter.rewriteColumnType(columnType, columnSize);
 				sb.append(columnTypeString);
 			} else {
 				sb.append(nativeType);
-			}
-			final Integer columnSize = column.getColumnSize();
-			if (columnSize != null) {
-				sb.append('(');
-				sb.append(columnSize.intValue());
-				sb.append(')');
+				if (columnSize != null) {
+				    sb.append('(');
+				    sb.append(columnSize.intValue());
+				    sb.append(')');
+				}
 			}
 			if (column.isNullable() != null && !column.isNullable().booleanValue()) {
 				sb.append(" NOT NULL");