+++ jdbc/src/main/java/org/apache/metamodel/jdbc/dialects/AbstractQueryRewriter.java
@@ -96,8 +96,19 @@ public abstract class AbstractQueryRewriter implements IQueryRewriter {
     }
 
     @Override
-    public String rewriteColumnType(ColumnType columnType) {
-        return columnType.toString();
+    public String rewriteColumnType(ColumnType columnType, Integer columnSize) {
+        return rewriteColumnTypeInternal(columnType.toString(), columnSize);
+    }
+    
+    protected String rewriteColumnTypeInternal(String columnType, Object columnParameter) {
+        final StringBuilder sb = new StringBuilder();
+        sb.append(columnType);
+        if (columnParameter != null) {
+            sb.append('(');
+            sb.append(columnParameter);
+            sb.append(')');
+        }
+        return sb.toString();
     }
 
     protected String rewriteOrderByClause(Query query, OrderByClause orderByClause) {