+++ jdbc/src/test/java/org/apache/metamodel/jdbc/integrationtests/OracleTest.java
@@ -28,6 +28,8 @@ import org.apache.metamodel.data.DataSet;
 import org.apache.metamodel.data.DataSetTableModel;
 import org.apache.metamodel.jdbc.JdbcDataContext;
 import org.apache.metamodel.jdbc.JdbcTestTemplates;
+import org.apache.metamodel.jdbc.dialects.IQueryRewriter;
+import org.apache.metamodel.jdbc.dialects.OracleQueryRewriter;
 import org.apache.metamodel.query.FromItem;
 import org.apache.metamodel.query.JoinType;
 import org.apache.metamodel.query.Query;
@@ -67,6 +69,15 @@ public class OracleTest extends AbstractJdbIntegrationTest {
     @Override
     protected String getPropertyPrefix() {
         return "oracle";
+    }
+    
+    public void testGetQueryRewriter() throws Exception {
+        if (!isConfigured()) {
+            return;
+        }
+
+        IQueryRewriter queryRewriter = getDataContext().getQueryRewriter();
+        assertEquals(OracleQueryRewriter.class, queryRewriter.getClass());
     }
 
     public void testCreateInsertAndUpdate() throws Exception {